+++ crates/forge_domain/src/env.rs
@@ -33,7 +33,7 @@ pub struct Environment {
     /// Configuration for the retry mechanism
     pub retry_config: RetryConfig,
     /// The maximum number of lines returned for FSSearch.
-    pub max_search_lines: u64,
+    pub max_search_lines: usize,
     /// Maximum bytes allowed for search results
     pub max_search_result_bytes: usize,
     /// Maximum characters for fetch content
